split networkd
