+++ zinit/boot.yaml
@@ -8,5 +8,5 @@ after:
   # list of all services that must be running before moving on
   - node-ready
   - storaged
-  - networkd
+  - initnet
   - identityd
